backend/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.11
# by the following command:
#
#    pip-compile --output-file=requirements.txt requirements.in
#
aiosqlite==0.21.0
    # via
    #   -r requirements.in
    #   databases
annotated-types==0.7.0
    # via pydantic
anyio==4.9.0
    # via
    #   httpx
    #   openai
    #   starlette
    #   watchfiles
asyncpg==0.30.0
    # via -r requirements.in
attrs==25.3.0
    # via
    #   jsonschema
    #   referencing
backoff==2.2.1
    # via posthog
certifi==2025.7.9
    # via
    #   httpcore
    #   httpx
    #   requests
charset-normalizer==3.4.2
    # via requests
click==8.2.1
    # via
    #   nltk
    #   uvicorn
databases[aiosqlite]==0.9.0
    # via -r requirements.in
distro==1.9.0
    # via
    #   openai
    #   posthog
docstring-parser==0.16
    # via haystack-ai
fastapi==0.116.0
    # via -r requirements.in
filelock==3.18.0
    # via
    #   huggingface-hub
    #   torch
    #   transformers
filetype==1.2.0
    # via haystack-experimental
fsspec==2025.5.1
    # via
    #   huggingface-hub
    #   torch
greenlet==3.2.3
    # via sqlalchemy
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
haystack-ai==2.15.2
    # via
    #   -r requirements.in
    #   haystack-experimental
    #   ollama-haystack
    #   pgvector-haystack
haystack-experimental==0.11.0
    # via haystack-ai
hf-xet==1.1.5
    # via huggingface-hub
httpcore==1.0.9
    # via httpx
httptools==0.6.4
    # via uvicorn
httpx==0.28.1
    # via
    #   -r requirements.in
    #   ollama
    #   openai
huggingface-hub==0.33.3
    # via
    #   sentence-transformers
    #   tokenizers
    #   transformers
idna==3.10
    # via
    #   anyio
    #   httpx
    #   requests
jinja2==3.1.6
    # via
    #   haystack-ai
    #   torch
jiter==0.10.0
    # via openai
joblib==1.5.1
    # via
    #   nltk
    #   scikit-learn
jsonschema==4.24.0
    # via haystack-ai
jsonschema-specifications==2025.4.1
    # via jsonschema
lazy-imports==1.0.0
    # via haystack-ai
markupsafe==3.0.2
    # via jinja2
more-itertools==10.7.0
    # via haystack-ai
mpmath==1.3.0
    # via sympy
networkx==3.5
    # via
    #   haystack-ai
    #   torch
nltk==3.9.1
    # via -r requirements.in
numpy==1.26.4
    # via
    #   -r requirements.in
    #   haystack-ai
    #   pgvector
    #   scikit-learn
    #   scipy
    #   sentence-transformers
    #   transformers
nvidia-cublas-cu12==12.6.4.1
    # via
    #   nvidia-cudnn-cu12
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cuda-cupti-cu12==12.6.80
    # via torch
nvidia-cuda-nvrtc-cu12==12.6.77
    # via torch
nvidia-cuda-runtime-cu12==12.6.77
    # via torch
nvidia-cudnn-cu12==9.5.1.17
    # via torch
nvidia-cufft-cu12==11.3.0.4
    # via torch
nvidia-cufile-cu12==1.11.1.6
    # via torch
nvidia-curand-cu12==10.3.7.77
    # via torch
nvidia-cusolver-cu12==11.7.1.2
    # via torch
nvidia-cusparse-cu12==12.5.4.2
    # via
    #   nvidia-cusolver-cu12
    #   torch
nvidia-cusparselt-cu12==0.6.3
    # via torch
nvidia-nccl-cu12==2.26.2
    # via torch
nvidia-nvjitlink-cu12==12.6.85
    # via
    #   nvidia-cufft-cu12
    #   nvidia-cusolver-cu12
    #   nvidia-cusparse-cu12
    #   torch
nvidia-nvtx-cu12==12.6.77
    # via torch
ollama==0.5.1
    # via ollama-haystack
ollama-haystack==1.1.0
    # via -r requirements.in
openai==1.95.0
    # via haystack-ai
packaging==25.0
    # via
    #   huggingface-hub
    #   transformers
pgvector==0.4.1
    # via pgvector-haystack
pgvector-haystack==3.4.0
    # via -r requirements.in
pillow==11.3.0
    # via sentence-transformers
posthog==5.4.0
    # via haystack-ai
psycopg[binary,pool]==3.2.9
    # via
    #   -r requirements.in
    #   pgvector-haystack
psycopg-binary==3.2.9
    # via psycopg
psycopg-pool==3.2.6
    # via psycopg
pydantic==2.11.7
    # via
    #   -r requirements.in
    #   fastapi
    #   haystack-ai
    #   ollama
    #   openai
    #   pydantic-settings
pydantic-core==2.33.2
    # via pydantic
pydantic-settings==2.10.1
    # via -r requirements.in
python-dateutil==2.9.0.post0
    # via
    #   haystack-ai
    #   posthog
python-dotenv==1.1.1
    # via
    #   -r requirements.in
    #   pydantic-settings
    #   uvicorn
python-multipart==0.0.20
    # via -r requirements.in
pyyaml==6.0.2
    # via
    #   haystack-ai
    #   huggingface-hub
    #   transformers
    #   uvicorn
referencing==0.36.2
    # via
    #   jsonschema
    #   jsonschema-specifications
regex==2024.11.6
    # via
    #   nltk
    #   transformers
requests==2.32.4
    # via
    #   haystack-ai
    #   huggingface-hub
    #   posthog
    #   transformers
rpds-py==0.26.0
    # via
    #   jsonschema
    #   referencing
safetensors==0.5.3
    # via transformers
scikit-learn==1.7.0
    # via
    #   -r requirements.in
    #   sentence-transformers
scipy==1.16.0
    # via
    #   scikit-learn
    #   sentence-transformers
sentence-transformers==2.7.0
    # via -r requirements.in
six==1.17.0
    # via
    #   posthog
    #   python-dateutil
sniffio==1.3.1
    # via
    #   anyio
    #   openai
sqlalchemy==2.0.41
    # via databases
starlette==0.46.2
    # via fastapi
sympy==1.14.0
    # via torch
tenacity==9.1.2
    # via haystack-ai
threadpoolctl==3.6.0
    # via scikit-learn
tokenizers==0.21.2
    # via transformers
torch==2.7.1
    # via
    #   -r requirements.in
    #   sentence-transformers
tqdm==4.67.1
    # via
    #   haystack-ai
    #   huggingface-hub
    #   nltk
    #   openai
    #   sentence-transformers
    #   transformers
transformers==4.53.2
    # via
    #   -r requirements.in
    #   sentence-transformers
triton==3.3.1
    # via torch
typing-extensions==4.14.1
    # via
    #   aiosqlite
    #   anyio
    #   fastapi
    #   haystack-ai
    #   huggingface-hub
    #   openai
    #   psycopg
    #   psycopg-pool
    #   pydantic
    #   pydantic-core
    #   referencing
    #   sqlalchemy
    #   torch
    #   typing-inspection
typing-inspection==0.4.1
    # via
    #   pydantic
    #   pydantic-settings
urllib3==2.5.0
    # via requests
uvicorn[standard]==0.35.0
    # via -r requirements.in
uvloop==0.21.0
    # via uvicorn
watchfiles==1.1.0
    # via uvicorn
websockets==15.0.1
    # via uvicorn

# The following packages are considered to be unsafe in a requirements file:
# setuptools
